pyproject.toml
[project]
name = "bathroom-button"
version = "0.1.0"
description = ""
authors = [
    {name = "momonala", email = "tinybathroom@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "RPi.GPIO; sys_platform == 'linux'",
    "httpx==0.23.3",
    "ipython",
    "paho-mqtt",
    "python-vlc",
    "requests",
    "youtube-search-python",
    "yt-dlp",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black",
    "jupyter",
    "notebook",
    "pandas>=2.3.3",
    "ruff>=0.14.10",
    "typer>=0.9.0",
    "isort>=7.0.0",
]

[tool.config]
# This project is primarily configured via secrets in src/values.py
# No non-secret configuration needed currently

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
download = "src.download_songs:main"
player = "src.player:main"
config = "src.config:main"

[tool.black]
line-length = 110
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
force_single_line = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-s -v --cov=src --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", "*/test_*.py"]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
